Restructured processEvent function in vl_databaselogger.
This function was and still is a mess. I restructured it for more
clarity. Furthermore, RPCs where not working because zera-modulemangaer
implements its owen processEvent function for this entity in ZeraDbLogger class.
To fix this I had to check, if RPCs are accepted. And remove vein RPC error response from
zeraDbLogger. Vl_databaselogger handles all error responses now.

Signed-off-by: bhamacher <b.hamacher@zera.de>