Add function to read session-> valuemap component.
Calling readSessionComponent it is possible to read one specific
componenten bound to the given session from valuemap.

Only components with a realtionship to session can be read.
The relationship is stored in sessions_valuemap.

The function will always return the last row.
That is equal to the latest and greates value.

Signed-off-by: bhamacher <b.hamacher@zera.de>