Replaced old VeinInterface with new vfcpp interface.
This is a huge commit because the complete API is changed. It will not
be possible to understand the changes by readign the code. Therefore all
changes will be listed here.

The changes in the Cpp and Vein API are in the same commit because the
VEIN API changes are needed, because the now used VfCpp API does not
support the old approach very well.

- The complete VeinApi is based on vfCpp now.

- Start logging is not done via component anymore but via Rpc now:
rpcName: RPC_startLogging
parameters: {"p_transaction" : "<transactionName>", "p_logType" : "<0/1/2>" , "p_duration" : "<durationInMs>", "p_guiContext" : "<guiContext>" ,"p_contentSets" : [<contentSets>]}
p_logType is a enumartion: 0:snapshot, 1:startStop 2:duration
The enum is provided by the ENUM_LogType component as QVariantMap
The duration parameter will only have effect in case Logtype is
duration.
Return: <transactionId> as int or -1 on fail.
At the moment the rpc will return -1 if the sessionName is not set or
logging is already enabled. Please note! It is possible to take
snapshots during logging.

- Stop logging is not done via component anymore but via Rpc now:
rpcName: stopLogging
parameters: {"p_transaction" : <transactionIdList>}
the rpc will stop all logging operations in the transactionIdList or all
logging operations, if the list is empty.

-Active trasaction are listed in the ActiveTransactions component.

-The Logging enabled component is readOnly now.

-The LoggendComponents component does not exist anymore.

-This entity has a proxy component now. It mirrors _System::Session.
 In case it changes all logging operations will be stopped.

Signed-off-by: bhamacher <b.hamacher@zera.de>