testloggersystem: Add 'feedEventLoop' when some changes to database are expected
In production environment, SQLiteDB runs on a separate thread. To get closer to this behavior, the signal/slot connection between DatabaseLogger and SQLiteDB is of type Qt::QueuedConnection.
But the test runs in a single thread and because of QueuedConnection the slots are not invoked immediately instead when the control returns to event loop.
So, we add extra 'feedEventLoop' in 'testloggersystem' when this asynchronous execution of slots is anticipated to make sure all events are processed.

Signed-off-by: amutalik <a.mutalik@zera.de>